Ensuring the main program can establish a ping with Elasticsearch
There was a minor bug in the program wherein the application didn't wait for Elasticsearch to connect completely. We compensated for this problem by having multiple restarts of the web service from docker until, eventually, the Elasticsearch connection was established.
My change has a segment of the code that ensures a healthy connection before proceeding.